add timestamp generator
