+++ .gitignore
@@ -1,3 +1,4 @@
 venv-pycord
 sakamata-pycord.code-workspace
 .env
+test.py
